`Reflip#flip`: Remove `request.features`.
Decorating the `request` object causes issues if a Reflip-enabled
Express app mounts a Reflip-enabled sub-app with different settings.
For example, if app A configures Reflip with the `"aardvarks"` setting
enabled, app B configures an independent instance of Reflip with
`"aardvarks"` disabled, and A mounts B as a sub-app,
`request.features.aardvarks` will be `false` if A mounts B after
mounting the `flip` middleware, and `true` if A mounts `flip` after
B.

Instead of adding a separate `exportFeaturesName` option,
`request.check()` now returns an object of feature names and their
corresponding states if it is called without any arguments.

This is a backward-incompatible change.